+++ src/components/menu/submenu-block/definition.js
@@ -1,0 +1,9 @@
+import SubmenuBlock from './';
+import Definition from './../../../../demo2/utils/definition';
+
+let definition = new Definition('submenu-block', SubmenuBlock, {
+  description: '[content needed] Basic example of the component',
+  designerNotes: '[content needed] Basic designs description for the component',
+});
+
+export default definition;