+++ src/components/mount-in-app/__spec__.js
@@ -44,7 +44,12 @@ describe('MountInApp', () => {
       expect(contentDiv.tagName).toEqual('DIV');
       expect(contentDiv.textContent).toEqual('Hello!!');
     });
+
+    describe('componentWillUnmount', () => {
+      it('removes all children', () => {
+        instance.unmount();
+        expect(placeholder.children.length).toEqual(0);
+      });
+    });
   });
-
 });
-